# docs/swagger.yaml
definitions:
...
  handler.OtherInStockItemReq:
    properties:
      price:
        type: number
      productId:
        type: integer
      quantity:
        type: number
      remark:
        type: string
    required:
    - productId
    - quantity
    type: object
  handler.OtherInStockSaveReq:
    properties:
      billDate:
        type: string
      complete:
        description: true=保存并过账
        type: boolean
      counterpart:
        type: string
      deptId:
        type: integer
      handlerId:
        type: integer
      inType:
        type: string
      items:
        items:
          $ref: '#/definitions/handler.OtherInStockItemReq'
        minItems: 1
        type: array
      remark:
        type: string
      settleUnit:
        type: string
      warehouseId:
        type: integer
    required:
    - billDate
    - inType
    - items
    - warehouseId
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /other-in-stocks:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 单号
        in: query
        name: keyword
        type: string
      - description: 商品名称/编号
        in: query
        name: productKw
        type: string
      - description: 入库仓库
        in: query
        name: warehouseId
        type: integer
      - description: 状态 draft/completed
        in: query
        name: status
        type: string
      - description: 入库类型
        in: query
        name: inType
        type: string
      - description: 录单时间起 YYYY-MM-DD
        in: query
        name: startDate
        type: string
      - description: 录单时间止 YYYY-MM-DD
        in: query
        name: endDate
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 其他入库单列表
      tags:
      - 库存
    post:
      parameters:
      - description: 单据内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.OtherInStockSaveReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增其他入库单（草稿或保存并过账）
      tags:
      - 库存
  /other-in-stocks/{id}:
    delete:
      parameters:
      - description: 单据ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除其他入库单
      tags:
      - 库存
    get:
      parameters:
      - description: 单据ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 其他入库单详情
      tags:
      - 库存
    put:
      parameters:
      - description: 单据ID
        in: path
        name: id
        required: true
        type: integer
      - description: 单据内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.OtherInStockSaveReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 编辑其他入库单
      tags:
      - 库存
  /other-in-stocks/{id}/complete:
    put:
      parameters:
      - description: 单据ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 其他入库单过账
      tags:
      - 库存
...
swagger: "2.0"